ML_Backend/requirements.txt
opencv-python
ultralytics
tensorflow
google-generativeai
pypdf
scikit-learn
Flask
chromadb